+++ src/index.ts
@@ -6,6 +6,8 @@ import StCheckbox from './components/checkbox/index.vue';
 import StCol from './components/col/index.vue';
 import StIcon from './components/icon/index.vue';
 import StInput from './components/input/index.vue';
+import { StNotificationService } from './components/notification/service';
+import StNotificationsGroup from './components/notifications-group/index.vue';
 import StPopover from './components/popover/index.vue';
 import StRow from './components/row/index.vue';
 import StTextarea from './components/textarea/index.vue';
@@ -17,18 +19,25 @@ const install = (Vue: VueConstructor) => {
   Vue.component('st-col', StCol);
   Vue.component('st-icon', StIcon);
   Vue.component('st-input', StInput);
+  Vue.component('st-notifications-group', StNotificationsGroup);
   Vue.component('st-popover', StPopover);
   Vue.component('st-row', StRow);
   Vue.component('st-textarea', StTextarea);
+
+  /* eslint-disable-next-line no-param-reassign */
+  Vue.prototype.$stNotification = StNotificationService;
 };
 
 export default {
   install,
   StButton,
   StButtonGroup,
+  StCheckbox,
+  StCol,
   StIcon,
   StInput,
+  StNotificationsGroup,
   StPopover,
   StRow,
-  StCol,
+  StTextarea,
 };