+++ stories/templates/notifications/default.template.js
@@ -1,0 +1,6 @@
+export const template = `
+<div>
+  <st-button type="primary" @click="callNotification" >Call notification</st-button>
+  <st-notifications-group :position="position" />
+</div>
+`;