+++ CHANGELOG.md
@@ -4,14 +4,13 @@
 
 ### Features
 * Added `Radio` component to `Offers` theme
+* Added new component: `Notifications`
+* Added new component: `Scrollbar`
 
 ### Maintenance
 * Fixed `Offers` theme for `Button` component
 * Fixed `Checkbox` value type 
 * Added components generator. If you are about to make a new component just run `npm run generator:component` and follow the instructions.
-
-### Features
-* Added `Scrollbar` component